+++ Sources/SpotifyExampleContent/ExamplePlaylists.swift
@@ -35,26 +35,6 @@ public extension PagingObject where Item == PlaylistItemContainer<Track> {
 
 }
 
-public extension Playlist where Items == PlaylistItems {
-    
-    /// Sample data for testing purposes.
-    ///
-    /// This playlist contains local tracks and episodes. Local tracks
-    /// often have most of their properties set to `nil`, which can be very
-    /// useful for testing purposes.
-    static let localSongs = Bundle.module.decodeJson(
-        // Playlist<PagingObject<PlaylistItemContainer<PlaylistItem>>>
-        forResource: "Local Songs - Playlist<PagingObject<PlaylistItemContainer<PlaylistItem>>>",
-        type: Self.self
-    )!
-    
-    /// Sample data for testing purposes.
-    static let crumb = Bundle.module.decodeJson(
-        forResource: "Crumb - Playlist<PagingObject<PlaylistItemContainer<PlaylistItem>>>",
-        type: Self.self
-    )!
-    
-}
 
 public extension PagingObject where Item == PlaylistItemContainer<PlaylistItem> {
     
@@ -67,6 +47,25 @@ public extension PagingObject where Item == PlaylistItemContainer<PlaylistItem> 
 
 }
 
+public extension Playlist where Items == PlaylistItems {
+    
+    /// Sample data for testing purposes.
+    ///
+    /// This playlist episodes and local tracks. Local tracks
+    /// often have most of their properties set to `nil`, which can be very
+    /// useful for testing purposes.
+    static let episodesAndLocalTracks = Bundle.module.decodeJson(
+        forResource: "Local Songs - Playlist<PagingObject<PlaylistItemContainer<PlaylistItem>>>",
+        type: Self.self
+    )!
+    
+    /// Sample data for testing purposes.
+    static let crumb = Bundle.module.decodeJson(
+        forResource: "Crumb - Playlist<PagingObject<PlaylistItemContainer<PlaylistItem>>>",
+        type: Self.self
+    )!
+    
+}
 
 public extension Playlist where Items == PlaylistsItemsReference {