+++ Sources/SpotifyWebAPI/API/SpotifyAPI+Episodes.swift
@@ -79,11 +79,12 @@ public extension SpotifyAPI {
        - market: *Optional*. An [ISO 3166-1 alpha-2 country code][2].
              If a country code is specified, only episodes that are available
              in that market will be returned. If the access token was granted
-             on behalf of a user (i.e., if you authorized your application
-             using the client credentials flow), the country associated with
-             the user account will take priority over this parameter. Users
-             can view the country that is associated with their account in
-             the [account settings][3].
+             on behalf of a user (i.e., if you authorized your application using
+             the authorization code flow or the authorization code flow with
+             proof key for code exchange), the country associated with the
+             user account will take priority over this parameter. Users can
+             view the country that is associated with their account in the
+             [account settings][3].
              
             **Note: If neither market or user country are provided, the**
              **episodes are considered unavailable for the client and**
@@ -96,7 +97,10 @@ public extension SpotifyAPI {
            or is unavailable in the specified market/country, `nil` is
            returned in the appropriate positions. Duplicate episode URIs in
            the request will result in duplicate episodes in the response.
-     
+           **Unlike many of the other endpoints for retrieving multiple**
+           **objects, if one of the URIs is invalid, then the entire request**
+           **will fail with a 400 "invalid id" error.**
+
      [1]: https://developer.spotify.com/documentation/web-api/reference/episodes/get-several-episodes/
      [2]: https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2
      [3]: https://www.spotify.com/account/overview/