+++ CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.3] -2020-10-16
+## [0.7.4] - 2020-10-17
+
+### Fixed
+
+* Fixed bugs for the `SpotifyAPI.showEpisodes(_:market:offset:limit:)` endpoint, which was trying to decode the wrong response type and did not add the id of the show to the query string correctly.
+
+### Added
+
+* Added `totalEpisodes` property to `Show`, which should've already been added.
+* Added example `PlaylistItem`s to the `SpotifyExampleContent` module.
+
+## [0.7.3] - 2020-10-16
 
 ### Changed
 
@@ -22,7 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-* Added documentation about how omitting the market parameter when using the client credentials flow causes episodes and shqows to not be returned.
+* Added documentation about how omitting the market parameter when using the client credentials flow causes episodes and shows to not be returned.
 * Added `SpotifyAPI.filteredPlaylistItems(_:filters:additionalTypes:limit:offset:market:)`.
 * Added `snapshotId` parameter to `SpotifyAPI.removeAllOccurencesFromPlaylist(_:of:snapshotId:)`.
 * Added `SpotifyPlayerError`; this error object is returned by Spotify when there are errors related to the player endpoints.